Feature/update story endpoint

+++ src/controllers/stories.js
@@ -25,30 +25,40 @@ export const createStoryController = async (req, res) => {
 export const patchStoryController = async (req, res) => {
   const { storyId } = req.params;
   const ownerId = req.user._id;
-  const updateFields = req.body;
+  const updateFields = req.body || {};
   const storyImageFile = req.file;
 
-  const { error } = updateStorySchema.validate(updateFields);
-  if (error) {
-    if (storyImageFile) {
-      await fs.unlink(storyImageFile.path);
+  const hasTextFields = Object.keys(updateFields).length > 0;
+  const hasFile = !!storyImageFile;
+
+  if (!hasTextFields && !hasFile) {
+    throw createHttpError(
+      400,
+      'At least one field or file must be provided for update.',
+    );
+  }
+
+  if (hasTextFields) {
+    const { error } = updateStorySchema.validate(updateFields);
+    if (error) {
+      if (storyImageFile) {
+        await fs.unlink(storyImageFile.path);
+      }
+      throw createHttpError(400, error.details[0].message);
     }
-    throw createHttpError(400, error.details[0].message);
   }
 
   if (updateFields.category) {
     const categoryExists = await checkCategoryExists(updateFields.category);
     if (!categoryExists) {
+      if (storyImageFile) {
+        await fs.unlink(storyImageFile.path);
+      }
       throw createHttpError(
         400,
         `Category with ID ${updateFields.category} not found.`,
       );
     }
-  }
-
-  const hasUpdates = Object.keys(updateFields).length > 0 || storyImageFile;
-  if (!hasUpdates) {
-    throw createHttpError(400, 'There are no fields to update.');
   }
 
   const updatedStory = await updateStoryById(
